requirements.txt
pynput
openai
python-dotenv
